#149: Rename some users groups and rights files
